+++ __tests__/App-test.tsx
@@ -4,11 +4,11 @@
 
 import 'react-native';
 import React from 'react';
-import App from '../app';
+import { LoginPage } from '../pages/login-page';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
-  renderer.create(<App />);
+  renderer.create(<LoginPage componentId={'Component1'} />);
 });
