+++ pages/login-page.tsx
@@ -8,47 +8,18 @@ import {
   useColorScheme,
   View,
 } from 'react-native';
-import { Styles } from './styles';
-import { setValue } from './persistency';
-import { loginValidator } from './regex';
-import { client } from './services/apollo';
+import { Styles } from '../styles';
+import { setValue } from '../persistency';
+import { loginValidator } from '../regex';
+import { client } from '../services/apollo';
 import { useMutation } from '@apollo/client';
-import { loginMutationGQL } from './graphql/mutations';
+import { loginMutationGQL } from '../graphql/mutations';
 import { Colors } from 'react-native/Libraries/NewAppScreen';
-import React, {useState, type PropsWithChildren} from 'react';
-import { Navigation } from 'react-native-navigation';
+import React, { useState } from 'react';
+import { Navigation, NavigationComponentProps } from 'react-native-navigation';
+import { Section } from '../section';
 
-const Section: React.FC<
-  PropsWithChildren<{
-    title: string;
-  }>
-> = ({children, title}) => {
-  const isDarkMode = useColorScheme() === 'dark';
-  return (
-    <View style={Styles.sectionContainer}>
-      <Text
-        style={[
-          Styles.sectionTitle,
-          {
-            color: isDarkMode ? Colors.white : Colors.black,
-          },
-        ]}>
-        {title}
-      </Text>
-      <Text
-        style={[
-          Styles.sectionDescription,
-          {
-            color: isDarkMode ? Colors.light : Colors.dark,
-          },
-        ]}>
-        {children}
-      </Text>
-    </View>
-  );
-};
-
-const App = (props: NavigationComponentProps) => {
+export const LoginPage = (props: NavigationComponentProps) => {
   const isDarkMode = useColorScheme() === 'dark';
   const backgroundStyle = {
     backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
@@ -133,5 +104,3 @@ const App = (props: NavigationComponentProps) => {
     </SafeAreaView>
   );
 };
-
-export default App;
